# crates/server/src/api/league_entries.rs
//! League registration: `/api/v1/leagues/{id}/entries` and admin pairing of solo entries.
//!
//! No fees in this step, so a complete entry is `confirmed` straight away (the
//! `pending_payment` status is reserved for the payments step). Every change locks the league
//! row first, which serialises registration per league.

use axum::{Json, extract::State, http::StatusCode};
use courtpit_domain::Discipline;
use serde::Deserialize;
use sqlx::{Postgres, QueryBuilder};
use utoipa::ToSchema;
use uuid::Uuid;

use crate::{
    ApiError, ApiResult, AppState, TenantTx,
    auth::CurrentPlayer,
    extract::{ApiJson, ApiPath, ApiQuery},
    leagues::{
        self, LeagueRow, LeagueStatus,
        entries::{self, ENTRY_COLUMNS, EntryRow, EntryStatus, EntryView},
    },
    players::{self, Names},
};

/// Loads a league the caller can see, locked, and requires registration to be open now.
async fn open_league(
    state: &AppState,
    tx: &mut TenantTx,
    id: Uuid,
    player: &CurrentPlayer,
) -> ApiResult<LeagueRow> {
    let league = leagues::load(tx, id, true).await?;
    if league.published_at.is_none() && !player.role.is_admin() {
        return Err(ApiError::NotFound("league"));
    }
    if !league.registration_open(state.clock.now()) {
        return Err(ApiError::conflict(
            "registration for this league is not open",
        ));
    }
    Ok(league)
}

async fn respond(
    tx: &mut TenantTx,
    league: Uuid,
    id: Uuid,
    viewer: &CurrentPlayer,
) -> ApiResult<Json<EntryView>> {
    let entry = entries::load(tx, league, id).await?;
    let names = Names::load(tx, viewer, entry.player_ids.iter().copied()).await?;
    Ok(Json(EntryView::for_viewer(entry, viewer, &names)))
}

/// Body of `POST /leagues/{id}/entries`.
#[derive(Debug, Default, Clone, Copy, Deserialize, ToSchema)]
#[serde(deny_unknown_fields)]
pub struct Register {
    /// Doubles/mixed: invite this partner; the entry confirms when they accept.
    pub partner_id: Option<Uuid>,
    /// Doubles/mixed: list yourself for others (and admins) looking for a partner.
    #[serde(default)]
    pub looking_for_partner: bool,
}

/// Registers the caller. Singles entries are confirmed immediately; doubles and mixed
/// entries wait for the invited partner (or a pairing) as `pending_partner`.
#[utoipa::path(post, path = "/api/v1/leagues/{id}/entries", tag = "leagues",
    params(("id" = Uuid, Path)), request_body = Register, security(("bearer" = [])),
    responses((status = 201, body = EntryView), (status = 403, body = crate::error::ErrorBody),
        (status = 404, body = crate::error::ErrorBody),
        (status = 409, body = crate::error::ErrorBody),
        (status = 422, body = crate::error::ErrorBody)))]
pub async fn register(
    State(state): State<AppState>,
    player: CurrentPlayer,
    ApiPath(league_id): ApiPath<Uuid>,
    ApiJson(body): ApiJson<Register>,
) -> ApiResult<(StatusCode, Json<EntryView>)> {
    player.require_verified()?;
    let mut tx = player.tenant.begin(&state.db).await?;
    let league = open_league(&state, &mut tx, league_id, &player).await?;
    let discipline = league.discipline();
    entries::require_not_entered(&mut tx, league_id, player.id, "you are").await?;
    let status = match (discipline, body.partner_id, body.looking_for_partner) {
        (Discipline::Singles, None, false) => EntryStatus::Confirmed,
        (Discipline::Singles, _, _) => {
            return Err(ApiError::validation("singles entries have no partner"));
        }
        (_, None, false) => {
            return Err(ApiError::validation(
                "name a partner_id or set looking_for_partner",
            ));
        }
        (_, Some(partner), _) => {
            if partner == player.id {
                return Err(ApiError::validation("you cannot partner yourself"));
            }
            players::require_active_members(&mut tx, &[partner]).await?;
            entries::require_not_entered(&mut tx, league_id, partner, "your partner is").await?;
            entries::check_mixed(&mut tx, &player.tenant, discipline, &[player.id, partner])
                .await?;
            EntryStatus::PendingPartner
        }
        (_, None, true) => {
            entries::check_mixed(&mut tx, &player.tenant, discipline, &[player.id]).await?;
            EntryStatus::PendingPartner
        }
    };
    let id = Uuid::now_v7();
    let _ = sqlx::query(
        "INSERT INTO league_entries (id, community_id, league_id, player_ids, created_by, status,
            looking_for_partner, invited_partner_id)
         VALUES ($1, $2, $3, $4, $5, $6, $7, $8)",
    )
    .bind(id)
    .bind(tx.community_id())
    .bind(league_id)
    .bind(vec![player.id])
    .bind(player.id)
    .bind(status)
    .bind(body.looking_for_partner && body.partner_id.is_none())
    .bind(body.partner_id)
    .execute(&mut *tx)
    .await?;
    let res = respond(&mut tx, league_id, id, &player).await?;
    tx.commit().await?;
    Ok((StatusCode::CREATED, res))
}

/// Filters for `GET /leagues/{id}/entries`.
#[derive(Debug, Clone, Copy, Deserialize, utoipa::IntoParams)]
#[into_params(parameter_in = Query)]
pub struct EntryQuery {
    /// Defaults to every status except `withdrawn`.
    pub status: Option<EntryStatus>,
    /// Only solo entries looking for a partner.
    #[serde(default)]
    pub looking_for_partner: bool,
}

/// A league's entries, in registration order.
#[utoipa::path(get, path = "/api/v1/leagues/{id}/entries", tag = "leagues",
    params(("id" = Uuid, Path), EntryQuery), security(("bearer" = [])),
    responses((status = 200, body = Vec<EntryView>), (status = 404, body = crate::error::ErrorBody)))]
pub async fn list_entries(
    State(state): State<AppState>,
    player: CurrentPlayer,
    ApiPath(league_id): ApiPath<Uuid>,
    ApiQuery(query): ApiQuery<EntryQuery>,
) -> ApiResult<Json<Vec<EntryView>>> {
    let mut tx = player.tenant.begin(&state.db).await?;
    let league = leagues::load(&mut tx, league_id, false).await?;
    if league.published_at.is_none() && !player.role.is_admin() {
        return Err(ApiError::NotFound("league"));
    }
    let mut qb: QueryBuilder<'_, Postgres> = QueryBuilder::new(format!(
        "SELECT {ENTRY_COLUMNS} FROM league_entries WHERE community_id = "
    ));
    let _ = qb
        .push_bind(tx.community_id())
        .push(" AND league_id = ")
        .push_bind(league_id);
    let _ = match query.status {
        Some(status) => qb.push(" AND status = ").push_bind(status),
        None => qb.push(" AND status <> 'withdrawn'"),
    };
    if query.looking_for_partner {
        let _ = qb.push(" AND looking_for_partner AND cardinality(player_ids) = 1");
    }
    let _ = qb.push(" ORDER BY created_at, id");
    let rows: Vec<EntryRow> = qb.build_query_as().fetch_all(&mut *tx).await?;
    let players = rows
        .iter()
        .flat_map(|entry| entry.player_ids.iter().copied());
    let names = Names::load(&mut tx, &player, players).await?;
    tx.commit().await?;
    Ok(Json(
        rows.into_iter()
            .map(|entry| EntryView::for_viewer(entry, &player, &names))
            .collect(),
    ))
}

/// Completes a pending doubles/mixed entry with its second player and confirms it.
async fn complete(tx: &mut TenantTx, entry: &EntryRow, partner: Uuid) -> ApiResult<()> {
    let _ = sqlx::query(
        "UPDATE league_entries SET player_ids = $3, status = 'confirmed',
            looking_for_partner = false, invited_partner_id = NULL, updated_at = now()
         WHERE community_id = $1 AND id = $2",
    )
    .bind(tx.community_id())
    .bind(entry.id)
    .bind(vec![entry.created_by, partner])
    .execute(&mut **tx)
    .await?;
    Ok(())
}

/// Withdraws `player`'s own solo, still-pending entries in `league` (superseded when they
/// join someone else's entry).
async fn withdraw_solo(tx: &mut TenantTx, league: Uuid, player: Uuid) -> ApiResult<()> {
    let _ = sqlx::query(
        "UPDATE league_entries SET status = 'withdrawn', invited_partner_id = NULL,
            updated_at = now()
         WHERE community_id = $1 AND league_id = $2 AND status = 'pending_partner'
           AND player_ids = ARRAY[$3]::uuid[]",
    )
    .bind(tx.community_id())
    .bind(league)
    .bind(player)
    .execute(&mut **tx)
    .await?;
    Ok(())
}

/// Accepts a partner invitation; the entry is confirmed. Any solo entry of your own in the
/// league is withdrawn.
#[utoipa::path(post, path = "/api/v1/leagues/{id}/entries/{entry_id}/accept", tag = "leagues",
    params(("id" = Uuid, Path), ("entry_id" = Uuid, Path)), security(("bearer" = [])),
    responses((status = 200, body = EntryView), (status = 403, body = crate::error::ErrorBody),
        (status = 404, body = crate::error::ErrorBody),
        (status = 409, body = crate::error::ErrorBody),
        (status = 422, body = crate::error::ErrorBody)))]
pub async fn accept_invite(
    State(state): State<AppState>,
    player: CurrentPlayer,
    ApiPath((league_id, entry_id)): ApiPath<(Uuid, Uuid)>,
) -> ApiResult<Json<EntryView>> {
    player.require_verified()?;
    let mut tx = player.tenant.begin(&state.db).await?;
    let league = open_league(&state, &mut tx, league_id, &player).await?;
    let entry = entries::load(&mut tx, league_id, entry_id).await?;
    if entry.invited_partner_id != Some(player.id) {
        return Err(ApiError::forbidden("you were not invited to this entry"));
    }
    if entry.status != EntryStatus::PendingPartner {
        return Err(ApiError::conflict(
            "this entry is no longer waiting for a partner",
        ));
    }
    withdraw_solo(&mut tx, league_id, player.id).await?;
    entries::require_not_entered(&mut tx, league_id, player.id, "you are").await?;
    players::require_active_members(&mut tx, &[entry.created_by, player.id]).await?;
    let pair = [entry.created_by, player.id];
    entries::check_mixed(&mut tx, &player.tenant, league.discipline(), &pair).await?;
    complete(&mut tx, &entry, player.id).await?;
    let res = respond(&mut tx, league_id, entry_id, &player).await?;
    tx.commit().await?;
    Ok(res)
}

/// Declines a partner invitation. The entry stays pending for its creator.
#[utoipa::path(post, path = "/api/v1/leagues/{id}/entries/{entry_id}/decline", tag = "leagues",
    params(("id" = Uuid, Path), ("entry_id" = Uuid, Path)), security(("bearer" = [])),
    responses((status = 200, body = EntryView), (status = 403, body = crate::error::ErrorBody),
        (status = 404, body = crate::error::ErrorBody)))]
pub async fn decline_invite(
    State(state): State<AppState>,
    player: CurrentPlayer,
    ApiPath((league_id, entry_id)): ApiPath<(Uuid, Uuid)>,
) -> ApiResult<Json<EntryView>> {
    let mut tx = player.tenant.begin(&state.db).await?;
    let _ = leagues::load(&mut tx, league_id, true).await?;
    let entry = entries::load(&mut tx, league_id, entry_id).await?;
    if entry.invited_partner_id != Some(player.id) {
        return Err(ApiError::forbidden("you were not invited to this entry"));
    }
    let _ = sqlx::query(
        "UPDATE league_entries SET invited_partner_id = NULL, updated_at = now()
         WHERE community_id = $1 AND id = $2",
    )
    .bind(tx.community_id())
    .bind(entry_id)
    .execute(&mut *tx)
    .await?;
    let res = respond(&mut tx, league_id, entry_id, &player).await?;
    tx.commit().await?;
    Ok(res)
}

/// Withdraws an entry: its players while registration is open, admins until the league
/// finishes (matches already scheduled stay; admins award walkovers).
#[utoipa::path(post, path = "/api/v1/leagues/{id}/entries/{entry_id}/withdraw", tag = "leagues",
    params(("id" = Uuid, Path), ("entry_id" = Uuid, Path)), security(("bearer" = [])),
    responses((status = 200, body = EntryView), (status = 403, body = crate::error::ErrorBody),
        (status = 404, body = crate::error::ErrorBody),
        (status = 409, body = crate::error::ErrorBody)))]
pub async fn withdraw(
    State(state): State<AppState>,
    player: CurrentPlayer,
    ApiPath((league_id, entry_id)): ApiPath<(Uuid, Uuid)>,
) -> ApiResult<Json<EntryView>> {
    let mut tx = player.tenant.begin(&state.db).await?;
    let league = leagues::load(&mut tx, league_id, true).await?;
    let entry = entries::load(&mut tx, league_id, entry_id).await?;
    let member = entry.player_ids.contains(&player.id);
    if !member && !player.role.is_admin() {
        return Err(ApiError::forbidden("this is not your entry"));
    }
    if entry.status == EntryStatus::Withdrawn {
        return Err(ApiError::conflict("the entry is already withdrawn"));
    }
    let allowed = if player.role.is_admin() {
        !matches!(
            league.status,
            LeagueStatus::Finished | LeagueStatus::Cancelled
        )
    } else {
        league.registration_open(state.clock.now())
    };
    if !allowed {
        return Err(ApiError::conflict(
            "entries can only be withdrawn while registration is open (ask an admin)",
        ));
    }
    let _ = sqlx::query(
        "UPDATE league_entries SET status = 'withdrawn', invited_partner_id = NULL,
            updated_at = now()
         WHERE community_id = $1 AND id = $2",
    )
    .bind(tx.community_id())
    .bind(entry_id)
    .execute(&mut *tx)
    .await?;
    let res = respond(&mut tx, league_id, entry_id, &player).await?;
    tx.commit().await?;
    Ok(res)
}

/// Body of `POST /admin/leagues/{id}/pair`.
#[derive(Debug, Clone, Copy, Deserialize, ToSchema)]
#[serde(deny_unknown_fields)]
pub struct PairBody {
    /// Two solo, pending entries. The first keeps its id and becomes the pair's entry; the
    /// second is withdrawn.
    pub entry_ids: [Uuid; 2],
}

/// Pairs two solo doubles/mixed entries into one confirmed entry (admin, before the draw).
#[utoipa::path(post, path = "/api/v1/admin/leagues/{id}/pair", tag = "admin",
    params(("id" = Uuid, Path)), request_body = PairBody, security(("bearer" = [])),
    responses((status = 200, body = EntryView), (status = 403, body = crate::error::ErrorBody),
        (status = 404, body = crate::error::ErrorBody),
        (status = 409, body = crate::error::ErrorBody),
        (status = 422, body = crate::error::ErrorBody)))]
pub async fn pair_entries(
    State(state): State<AppState>,
    admin: CurrentPlayer,
    ApiPath(league_id): ApiPath<Uuid>,
    ApiJson(body): ApiJson<PairBody>,
) -> ApiResult<Json<EntryView>> {
    admin.require_admin()?;
    let [first, second] = body.entry_ids;
    if first == second {
        return Err(ApiError::validation("pair two different entries"));
    }
    let mut tx = admin.tenant.begin(&state.db).await?;
    let league = leagues::load(&mut tx, league_id, true).await?;
    if league.status != LeagueStatus::Registration {
        return Err(ApiError::conflict("pairing happens before the draw"));
    }
    let first_entry = entries::load(&mut tx, league_id, first).await?;
    let second_entry = entries::load(&mut tx, league_id, second).await?;
    for entry in [&first_entry, &second_entry] {
        if entry.status != EntryStatus::PendingPartner || entry.player_ids.len() != 1 {
            return Err(ApiError::conflict(
                "both entries must be solo and waiting for a partner",
            ));
        }
    }
    let pair = [first_entry.created_by, second_entry.created_by];
    players::require_active_members(&mut tx, &pair).await?;
    entries::check_mixed(&mut tx, &admin.tenant, league.discipline(), &pair).await?;
    let _ = sqlx::query(
        "UPDATE league_entries SET status = 'withdrawn', invited_partner_id = NULL,
            looking_for_partner = false, updated_at = now()
         WHERE community_id = $1 AND id = $2",
    )
    .bind(tx.community_id())
    .bind(second_entry.id)
    .execute(&mut *tx)
    .await?;
    complete(&mut tx, &first_entry, second_entry.created_by).await?;
    let res = respond(&mut tx, league_id, first_entry.id, &admin).await?;
    tx.commit().await?;
    Ok(res)
}