# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Courtpit API",
    "description": "Courtpit REST API.\n\n**Community.** Every request is scoped to one community. Native clients send `X-Courtpit-Community: <slug>`; browsers are resolved from the host (`<slug>.courtpit.app` or a registered custom domain). The header wins over the host. A request without a resolvable community answers `400`, an unknown one `404`.\n\n**Authentication.** A session is an opaque token, sent as `Authorization: Bearer <token>` (native) or as the `courtpit_session` cookie (web; ask for it at sign-in with `X-Courtpit-Client: web`). Either one authenticates.\n\n**Errors.** Every error is `{ \"error\": { \"code\", \"message\" } }`; clients switch on `code`.\n\n**Lists** are cursor-paginated: `?cursor=&limit=` in, `{ items, next_cursor }` out.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/leagues/{id}/pair": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Pairs two solo doubles/mixed entries into one confirmed entry (admin, before the draw).",
        "operationId": "pair_entries",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
...
    "/api/v1/leagues/{id}/entries": {
      "get": {
        "tags": [
          "leagues"
        ],
        "summary": "A league's entries, in registration order.",
        "operationId": "list_entries",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Defaults to every status except `withdrawn`.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EntryStatus"
            }
          },
          {
            "name": "looking_for_partner",
            "in": "query",
            "description": "Only solo entries looking for a partner.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntryView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "leagues"
        ],
        "summary": "Registers the caller. Singles entries are confirmed immediately; doubles and mixed\nentries wait for the invited partner (or a pairing) as `pending_partner`.",
        "operationId": "register",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/leagues/{id}/entries/{entry_id}/accept": {
      "post": {
        "tags": [
          "leagues"
        ],
        "summary": "Accepts a partner invitation; the entry is confirmed. Any solo entry of your own in the\nleague is withdrawn.",
        "operationId": "accept_invite",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/leagues/{id}/entries/{entry_id}/decline": {
      "post": {
        "tags": [
          "leagues"
        ],
        "summary": "Declines a partner invitation. The entry stays pending for its creator.",
        "operationId": "decline_invite",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/leagues/{id}/entries/{entry_id}/withdraw": {
      "post": {
        "tags": [
          "leagues"
        ],
        "summary": "Withdraws an entry: its players while registration is open, admins until the league\nfinishes (matches already scheduled stay; admins award walkovers).",
        "operationId": "withdraw",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntryStatus": {
        "type": "string",
        "description": "Where an entry stands.",
        "enum": [
          "pending_partner",
          "pending_payment",
          "confirmed",
          "withdrawn"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The JSON body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "What went wrong."
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "description": "Machine-readable code plus a human-readable message.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code; clients switch on it. One of `bad_request` (400),\n`validation_failed` (422), `unauthorized` and `invalid_credentials` (401), `forbidden`\n(403), `not_found` (404), `conflict` (409), `rate_limited` (429) or `internal` (500)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation; never contains internal details."
          }
        }
      },
...
      "PairBody": {
        "type": "object",
        "description": "Body of `POST /admin/leagues/{id}/pair`.",
        "required": [
          "entry_ids"
        ],
        "properties": {
          "entry_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Two solo, pending entries. The first keeps its id and becomes the pair's entry; the\nsecond is withdrawn."
          }
        },
        "additionalProperties": false
      },
...
      "Register": {
        "type": "object",
        "description": "Body of `POST /leagues/{id}/entries`.",
        "properties": {
          "looking_for_partner": {
            "type": "boolean",
            "description": "Doubles/mixed: list yourself for others (and admins) looking for a partner."
          },
          "partner_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Doubles/mixed: invite this partner; the entry confirms when they accept."
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token (native clients)."
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "courtpit_session",
        "description": "Session cookie (web clients); alternative to the bearer token."
      }
    }
  },
...
}